Added fileName

+++ src/main/java/org/arcticwolf/service/impl/PropertiesFilter.java
@@ -1,6 +1,8 @@
 package org.arcticwolf.service.impl;
 
 import static org.arcticwolf.config.ConfigProperties.config;
+
+import org.arcticwolf.model.CreatePropertiesRequest;
 import org.arcticwolf.service.Task;
 
 import java.io.File;
@@ -19,9 +21,11 @@ public class PropertiesFilter implements Task {
     @Override
     public void doWork(String... input) {
         System.out.println("Came to doWork "+Thread.currentThread().getName());
+        String fileName = "";
         String pathToFileCreated;
-        if (input.length == 1) {
-            pathToFileCreated = input[0];
+        if (input.length == 2) {
+            pathToFileCreated = input[0] + "/" + input[1];
+            fileName = input[1];
         } else {
             System.out.println("Input not suitable for this flow" +input.length);
             return;
@@ -46,9 +50,18 @@ public class PropertiesFilter implements Task {
 
         Map<String, String> properties = convertPropertiesToMap(prop);
         System.out.println(properties);
+
+        CreatePropertiesRequest createPropertiesRequest = new CreatePropertiesRequest();
+        createPropertiesRequest.setProperties(properties);
+        createPropertiesRequest.setFileName(fileName);
+
+        boolean sendSuccess = sendToServer(createPropertiesRequest);
+        if (!sendSuccess) {
+            System.out.println("Failed to send properties to server");
+            return;
+        }
         boolean deleted = deleteFile(pathToFileCreated);
         System.out.println("Deleted : "+deleted);
-        sendToServer(properties);
     }
 
     private Map<String, String> convertPropertiesToMap(Properties props) {
@@ -67,19 +80,20 @@ public class PropertiesFilter implements Task {
         return map;
     }
 
-    private void sendToServer(Map<String, String> map) {
+    private boolean sendToServer(CreatePropertiesRequest createPropertiesRequest) {
         String server = config.get("server.address");
         Integer port = Integer.parseInt(config.get("server.port"));
         try(
                 Socket socket = new Socket(server, port);
                 ObjectOutputStream out = new ObjectOutputStream(socket.getOutputStream());
                 ) {
-            out.writeObject(map);
+            out.writeObject(createPropertiesRequest);
 
         } catch (IOException e) {
             e.printStackTrace();
+            return false;
         }
-
+        return true;
     }
 
     private boolean deleteFile(String path) {
